move helper function files into helper dir
