+++ server/index.js
@@ -3,7 +3,7 @@ import { config } from 'dotenv';
 import cors from 'cors';
 import bodyParser from 'body-parser';
 import { addNewEntry, searchIt } from '../database/index';
-import addExperience from './addExperience';
+import addExperience from '../helpers/addExperience';
 
 config();
 const app = express();
@@ -25,10 +25,7 @@ app.get('/rentals?:location', (req, res) => {
     .catch(err => res.send(err));
 });
 
-app.post('/rentals', (req, res) => {
-  // TODO: remove res.send result/err after testing is done
-  const { location, index, body } = req.body;
-  addNewEntry(location, index, body)
-    .then(result => res.send(result))
-    .catch(err => res.send(err));
+app.post('/reservations', (req, res) => {
+  // TODO: should package up
+
 });
